requirements.txt
streamlit
numpy
pillow
requests
scikit-learn
joblib
